lab5-comunicação-indireta-pubsub-2/pyproject.toml
[tool.poetry]
name = "lab5-comunicação-indireta-pubsub-2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "lab5_comunicação_indireta_pubsub_2"}]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
